Modelos,Controladores Actualizados agregando funcionalidades

+++ IntDevPos/Controlador/Controlador_Principal.cs
@@ -4,6 +4,7 @@ using IntDevPos.Modelos_pos;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -31,9 +32,9 @@ namespace IntDevPos.Controlador
                     {
                         // Ruta nueva para Json Exitosos 
                         //string RutaNuevaI = CarpetaOrdenesInsertadas + Path.GetFileName(reg);
-                        string RutaNuevaI = CarpetaOrdenesInsertadas + "Procesado" + Path.GetFileName(reg);
+                        string RutaNuevaI = CarpetaOrdenesInsertadas + "Procesado-" + Path.GetFileName(reg);
                         // Ruta nueva para Json Erroneos 
-                        string RutaNuevaE = CarpetaOrdenesErroneas + "Erronea" + Path.GetFileName(reg);
+                        string RutaNuevaE = CarpetaOrdenesErroneas + "Erronea-" + Path.GetFileName(reg);
 
                         /* Se agrega trycatch para el deserializar el json
                          * Dado el caso sea exitoso copia la orden a la carpeta procesadas
@@ -51,16 +52,21 @@ namespace IntDevPos.Controlador
                             }
                             else
                             {
-                                File.Copy(reg, RutaNuevaE);
+                                if (!File.Exists(RutaNuevaE))
+                                {
+                                    File.Copy(reg, RutaNuevaE);
+                                }
                             }
-
                         }
-                        catch (Exception)
+                        catch (Exception )
                         {
-                            File.Copy(reg, RutaNuevaE);
+                            if (!File.Exists(RutaNuevaE))
+                            {
+                                File.Copy(reg, RutaNuevaE);
+                            } 
                         }
-
-
+                        // Elimina la orden independientemente fuera exitosa o erronea
+                        File.Delete(reg);
                     }
                 }
 
@@ -96,21 +102,38 @@ namespace IntDevPos.Controlador
         private bool RecorrerOrden(Principal modelPrin)
         {
             int contador = 0;
-            string IdDevInMotion = modelPrin.OrderreferenceId;
-            foreach (var pOrden in modelPrin.salesLinesList)
-            {
-                if (InsertarOrdenes(IdDevInMotion, pOrden.productId, pOrden.salesQty, pOrden.salesPrice, contador))
-                {
-                    contador++;
-                }
+            // Cabecera
+            string IdDevInMotion  = modelPrin.OrderreferenceId;
+            string Identification = modelPrin.customer.identification;
+            string ExpeditionDate = modelPrin.customer.expeditionDate;
+            string PrimerNombre   = modelPrin.customer.firstName;
+            string PrimerApellido = modelPrin.customer.secondName;
+            string Cumpleanos     = modelPrin.customer.birthDate;
+            string email          = modelPrin.customer.email;
+            string celular        = modelPrin.customer.phoneNumber;
+            string direccion      = modelPrin.customer.Address;
 
+            // Agrega la cabecera 
+            if (InsertarCab(IdDevInMotion,Identification,ExpeditionDate,PrimerNombre,PrimerApellido,Cumpleanos,email,celular,direccion))
+            {
+                // Recorre salesLinesList para agregar el detalle.
+                foreach (var pOrden in modelPrin.salesLinesList)
+                {
+                    // Detalle
+                    if (InsertarDet(IdDevInMotion, pOrden.productId, pOrden.salesQty, pOrden.salesPrice))
+                    {
+                        contador++;
+                    }
+                }
             }
+            
 
             return contador > 0 ? true : false;
         }
 
         public bool InsertarDet(string IdDev, string Codigo, int Cant, string Precio)
         {
+            bool isInsertDet = false;
             string sqlDetInt = "INSERT INTO " + Constantes_T.TBL_DETWEB + "(" +
                             Det_web.IdDevInMotion + "," + Det_web.Codigo + "," +
                             Det_web.Cantidad + "," + Det_web.Valor + "," +
@@ -120,22 +143,40 @@ namespace IntDevPos.Controlador
             // Hace conexion con la tabla detalle.
             if (con.Connect(Det_web.RutaTabla))
             {
-                con.ejecutar(sqlDetInt);
+                if (con.ejecutar(sqlDetInt)>0)
+                {
+                    isInsertDet = true;
+                }
+                
             }
+
+            return isInsertDet ;
         }
 
-        public bool InsertarCab(string IdDev)
+        public bool InsertarCab(string IdDev, string identification, string expeditionDate, string primerNombre, string primerApellido, string cumpleanos, string email, string celular, string direccion,string payterm, string observa)
         {
+            bool isInsertCab = false;
+
             string sqlCabInt = "INSERT INTO " + Constantes_T.TBL_CABWEB + "(" +
-                            Cab_web.IdDevInMotion + ") VALUES('" + IdDev + "')";
+                            Cab_web.IdDevInMotion + "," + Cab_web.Identification + "," +
+                            Cab_web.ExpeditionDate + ","+ Cab_web.FirstName + ","+ 
+                            Cab_web.SecondName + "," + Cab_web.BirthDay + "," + 
+                            Cab_web.Email + ","+ Cab_web.PhoneNumbrer + ","+
+                            Cab_web.Address + ","+ Cab_web.PayTerm + ","+Cab_web.Observations+
+                            ") VALUES('" + IdDev + "','"+ identification+"','"+ expeditionDate + "','"+
+                            primerNombre + "','"+ primerApellido + "','"+ cumpleanos+ "','"+ email + 
+                            "','"+ celular + "','"+ direccion + "','"+ payterm + "','"+ observa + "')";
 
             // Hace conexion con la tabla cabecera.
             if (con.Connect(Cab_web.RutaTabla))
             {
                 if (con.ejecutar(sqlCabInt) > 0)
                 {
+                    isInsertCab = true;
                 }
             }
+
+            return isInsertCab;
         }
 
     }
